backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.9.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.16.1
idna==3.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
twilio==9.8.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.11.0

# Azure deployment dependencies
gunicorn==21.2.0
azure-storage-blob==12.19.0

# Database management and migration tools
django-migration-testcase==0.0.15
django-dbbackup==4.1.0